Move usage handlers into platform core
