+++ addons/smart_core/handlers/usage_track.py
@@ -1,0 +1,128 @@
+# -*- coding: utf-8 -*-
+from __future__ import annotations
+
+import logging
+
+from odoo import fields
+from odoo.addons.smart_core.core.base_handler import BaseIntentHandler
+
+
+_logger = logging.getLogger(__name__)
+
+
+class UsageTrackHandler(BaseIntentHandler):
+    INTENT_TYPE = "usage.track"
+    DESCRIPTION = "Track scene/capability usage counters"
+    VERSION = "1.0.0"
+    ETAG_ENABLED = False
+    REQUIRED_GROUPS = ["base.group_user"]
+    ACL_MODE = "explicit_check"
+    NON_IDEMPOTENT_ALLOWED = "analytics counters are append-only metrics and intentionally non-replayable"
+    SOURCE_AUTHORITY = {
+        "kind": "usage_analytics_counter_write_proxy",
+        "authorities": ["sc.usage.counter", "res.groups", "odoo.orm"],
+        "projection_only": False,
+        "observability_only": True,
+        "no_business_fact_authority": True,
+        "write_authority": "sc.usage.counter.bump",
+    }
+
+    def _bump(self, usage_model, company, key):
+        if usage_model is None or not company or not key:
+            return
+        try:
+            usage_model.sudo().bump(company, key, 1)
+        except Exception as exc:
+            _logger.warning("[usage.track] bump failed company=%s key=%s error=%s", company.id, key, exc)
+
+    def _day_key(self):
+        return fields.Date.context_today(self.env.user).strftime("%Y-%m-%d")
+
+    def _role_codes_for_user(self, user):
+        if not user:
+            return []
+        role_codes = set()
+        group_xmlids = user.groups_id.get_external_id().values()
+        prefix = "smart_construction_core.group_sc_role_"
+        for xmlid in group_xmlids:
+            if isinstance(xmlid, str) and xmlid.startswith(prefix):
+                role_codes.add(xmlid[len(prefix):])
+        if user.has_group("base.group_system"):
+            role_codes.add("admin")
+        return sorted(role_codes)
+
+    def handle(self, payload=None, ctx=None):
+        params = payload or self.params or {}
+        if isinstance(params, dict) and isinstance(params.get("params"), dict):
+            # Intent router passes envelope: {intent, params, context}
+            params = params.get("params") or {}
+        event_type = str(params.get("event_type") or "").strip().lower()
+        scene_key = str(params.get("scene_key") or "").strip()
+        capability_key = str(params.get("capability_key") or "").strip()
+
+        user = self.env.user
+        company = user.company_id if user else None
+        Usage = self.env.get("sc.usage.counter")
+        role_codes = self._role_codes_for_user(user)
+        uid = int(user.id or 0) if user else 0
+
+        tracked = []
+        day_key = self._day_key()
+        if event_type == "scene_open" and scene_key:
+            self._bump(Usage, company, "usage.scene_open.total")
+            self._bump(Usage, company, f"usage.scene_open.{scene_key}")
+            self._bump(Usage, company, f"usage.scene_open.daily.{day_key}")
+            if uid:
+                self._bump(Usage, company, f"usage.scene_open.user.{uid}.total")
+                self._bump(Usage, company, f"usage.scene_open.user.{uid}.{scene_key}")
+                self._bump(Usage, company, f"usage.scene_open.user.{uid}.daily.{day_key}")
+                tracked.extend([
+                    f"usage.scene_open.user.{uid}.total",
+                    f"usage.scene_open.user.{uid}.{scene_key}",
+                    f"usage.scene_open.user.{uid}.daily.{day_key}",
+                ])
+            for role_code in role_codes:
+                self._bump(Usage, company, f"usage.scene_open.role.{role_code}.total")
+                self._bump(Usage, company, f"usage.scene_open.role.{role_code}.{scene_key}")
+                self._bump(Usage, company, f"usage.scene_open.role.{role_code}.daily.{day_key}")
+                tracked.extend([
+                    f"usage.scene_open.role.{role_code}.total",
+                    f"usage.scene_open.role.{role_code}.{scene_key}",
+                    f"usage.scene_open.role.{role_code}.daily.{day_key}",
+                ])
+            tracked.extend([
+                "usage.scene_open.total",
+                f"usage.scene_open.{scene_key}",
+                f"usage.scene_open.daily.{day_key}",
+            ])
+        elif event_type == "capability_open" and capability_key:
+            self._bump(Usage, company, "usage.capability_open.total")
+            self._bump(Usage, company, f"usage.capability_open.{capability_key}")
+            self._bump(Usage, company, f"usage.capability_open.daily.{day_key}")
+            if uid:
+                self._bump(Usage, company, f"usage.capability_open.user.{uid}.total")
+                self._bump(Usage, company, f"usage.capability_open.user.{uid}.{capability_key}")
+                self._bump(Usage, company, f"usage.capability_open.user.{uid}.daily.{day_key}")
+                tracked.extend([
+                    f"usage.capability_open.user.{uid}.total",
+                    f"usage.capability_open.user.{uid}.{capability_key}",
+                    f"usage.capability_open.user.{uid}.daily.{day_key}",
+                ])
+            for role_code in role_codes:
+                self._bump(Usage, company, f"usage.capability_open.role.{role_code}.total")
+                self._bump(Usage, company, f"usage.capability_open.role.{role_code}.{capability_key}")
+                self._bump(Usage, company, f"usage.capability_open.role.{role_code}.daily.{day_key}")
+                tracked.extend([
+                    f"usage.capability_open.role.{role_code}.total",
+                    f"usage.capability_open.role.{role_code}.{capability_key}",
+                    f"usage.capability_open.role.{role_code}.daily.{day_key}",
+                ])
+            tracked.extend([
+                "usage.capability_open.total",
+                f"usage.capability_open.{capability_key}",
+                f"usage.capability_open.daily.{day_key}",
+            ])
+        else:
+            return {"ok": False, "error": {"code": 400, "message": "invalid usage params"}, "meta": {"intent": self.INTENT_TYPE, "source_authority": self.SOURCE_AUTHORITY}}
+
+        return {"ok": True, "data": {"tracked": tracked, "event_type": event_type}, "meta": {"intent": self.INTENT_TYPE, "source_authority": self.SOURCE_AUTHORITY}}
