+++ docs/ops/audit/write_intent_permission_audit.md
@@ -6,7 +6,7 @@
 
 | intent | exists | required_groups | acl_guard | sudo_calls | unguarded_sudo | risk | source |
 |---|---:|---:|---:|---:|---:|---|---|
-| execute_button | Y | 1 | Y | 3 | 0 | low | addons/smart_core/handlers/execute_button.py |
+| execute_button | Y | 1 | Y | 5 | 0 | low | addons/smart_core/handlers/execute_button.py |
 - matched: `execute_button`
 - note: `execute_button` execute_button: permission posture looks acceptable
 | api.data(write) | Y | 1 | Y | 0 | 0 | low | addons/smart_core/handlers/api_data_batch.py, addons/smart_core/handlers/api_data_unlink.py, addons/smart_core/handlers/api_data_write.py |
@@ -20,7 +20,7 @@
 | file.upload | Y | 1 | Y | 0 | 0 | low | addons/smart_core/handlers/file_upload.py |
 - matched: `file.upload`
 - note: `file.upload` file.upload: permission posture looks acceptable
-| report.export | Y | 1 | Y | 1 | 0 | low | addons/smart_construction_core/handlers/usage_export_csv.py |
+| report.export | Y | 1 | Y | 1 | 0 | low | addons/smart_core/handlers/usage_export_csv.py |
 - matched: `usage.export.csv`
 - note: `report.export` usage.export.csv: permission posture looks acceptable
 | job.cancel | N | 0 | N | 0 | 0 | low | - |
